Fix race condition between load and frame ready
There was a race condition between these two operations. If the frame was ready before data had been given to the montage-frame through load() we could get in a state where we were reseting the frame application at the same time we were trying to set it up.

Now we make sure we don't try to setup the frame application before having data to do so.